requirements.txt
astroid==2.4.2
backcall==0.2.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
future==0.18.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.0
seaborn==0.10.1
six==1.15.0
toml==0.10.1
torch==1.5.1
torchvision==0.6.1
traitlets==4.3.3
wcwidth==0.2.5
wrapt==1.12.1
